requirements.txt
PyYAML==5.4.1
discord-components==1.0.0
discord.py==1.7.1
Flask==1.1.2